+++ pytorch_object_detection/faster_rcnn/backbone/__init__.py
@@ -1,3 +1,4 @@
 from .resnet50_fpn_model import resnet50_fpn_backbone
 from .mobilenetv2_model import MobileNetV2
 from .vgg_model import vgg
+from .feature_pyramid_network import LastLevelMaxPool, BackboneWithFPN
