+++ pytorch_object_detection/faster_rcnn/backbone/resnet50_fpn_model.py
@@ -1,19 +1,17 @@
 import os
-from collections import OrderedDict
 
 import torch
 import torch.nn as nn
-from torch.jit.annotations import List, Dict
 from torchvision.ops.misc import FrozenBatchNorm2d
 
-from .feature_pyramid_network import FeaturePyramidNetwork, LastLevelMaxPool
+from .feature_pyramid_network import BackboneWithFPN, LastLevelMaxPool
 
 
 class Bottleneck(nn.Module):
     expansion = 4
 
     def __init__(self, in_channel, out_channel, stride=1, downsample=None, norm_layer=None):
-        super(Bottleneck, self).__init__()
+        super().__init__()
         if norm_layer is None:
             norm_layer = nn.BatchNorm2d
 
@@ -56,7 +54,7 @@ class Bottleneck(nn.Module):
 class ResNet(nn.Module):
 
     def __init__(self, block, blocks_num, num_classes=1000, include_top=True, norm_layer=None):
-        super(ResNet, self).__init__()
+        super().__init__()
         if norm_layer is None:
             norm_layer = nn.BatchNorm2d
         self._norm_layer = norm_layer
@@ -134,100 +132,6 @@ def overwrite_eps(model, eps):
     for module in model.modules():
         if isinstance(module, FrozenBatchNorm2d):
             module.eps = eps
-
-
-class IntermediateLayerGetter(nn.ModuleDict):
-    """
-    Module wrapper that returns intermediate layers from a model
-    It has a strong assumption that the modules have been registered
-    into the model in the same order as they are used.
-    This means that one should **not** reuse the same nn.Module
-    twice in the forward if you want this to work.
-    Additionally, it is only able to query submodules that are directly
-    assigned to the model. So if `model` is passed, `model.feature1` can
-    be returned, but not `model.feature1.layer2`.
-    Arguments:
-        model (nn.Module): model on which we will extract the features
-        return_layers (Dict[name, new_name]): a dict containing the names
-            of the modules for which the activations will be returned as
-            the key of the dict, and the value of the dict is the name
-            of the returned activation (which the user can specify).
-    """
-    __annotations__ = {
-        "return_layers": Dict[str, str],
-    }
-
-    def __init__(self, model, return_layers):
-        if not set(return_layers).issubset([name for name, _ in model.named_children()]):
-            raise ValueError("return_layers are not present in model")
-
-        orig_return_layers = return_layers
-        return_layers = {str(k): str(v) for k, v in return_layers.items()}
-        layers = OrderedDict()
-
-        # 遍历模型子模块按顺序存入有序字典
-        # 只保存layer4及其之前的结构，舍去之后不用的结构
-        for name, module in model.named_children():
-            layers[name] = module
-            if name in return_layers:
-                del return_layers[name]
-            if not return_layers:
-                break
-
-        super(IntermediateLayerGetter, self).__init__(layers)
-        self.return_layers = orig_return_layers
-
-    def forward(self, x):
-        out = OrderedDict()
-        # 依次遍历模型的所有子模块，并进行正向传播，
-        # 收集layer1, layer2, layer3, layer4的输出
-        for name, module in self.items():
-            x = module(x)
-            if name in self.return_layers:
-                out_name = self.return_layers[name]
-                out[out_name] = x
-        return out
-
-
-class BackboneWithFPN(nn.Module):
-    """
-    Adds a FPN on top of a model.
-    Internally, it uses torchvision.models._utils.IntermediateLayerGetter to
-    extract a submodel that returns the feature maps specified in return_layers.
-    The same limitations of IntermediatLayerGetter apply here.
-    Arguments:
-        backbone (nn.Module)
-        return_layers (Dict[name, new_name]): a dict containing the names
-            of the modules for which the activations will be returned as
-            the key of the dict, and the value of the dict is the name
-            of the returned activation (which the user can specify).
-        in_channels_list (List[int]): number of channels for each feature map
-            that is returned, in the order they are present in the OrderedDict
-        out_channels (int): number of channels in the FPN.
-        extra_blocks: ExtraFPNBlock
-    Attributes:
-        out_channels (int): the number of channels in the FPN
-    """
-
-    def __init__(self, backbone, return_layers, in_channels_list, out_channels, extra_blocks=None):
-        super(BackboneWithFPN, self).__init__()
-
-        if extra_blocks is None:
-            extra_blocks = LastLevelMaxPool()
-
-        self.body = IntermediateLayerGetter(backbone, return_layers=return_layers)
-        self.fpn = FeaturePyramidNetwork(
-            in_channels_list=in_channels_list,
-            out_channels=out_channels,
-            extra_blocks=extra_blocks,
-        )
-
-        self.out_channels = out_channels
-
-    def forward(self, x):
-        x = self.body(x)
-        x = self.fpn(x)
-        return x
 
 
 def resnet50_fpn_backbone(pretrain_path="",
